Commonise maxwell3d texture code
This will be shared with the compute engine implementation.
